// src/formats/png/png-color-type.ts
/** @format */

/**
 * Enum representing PNG color types.
 */
export enum PngColorType {
  /**
   * Grayscale color type.
   */
  grayscale = 0,

  /**
   * RGB color type.
   */
  rgb = 2,

  /**
   * Indexed color type.
   */
  indexed = 3,

  /**
   * Grayscale with alpha channel color type.
   */
  grayscaleAlpha = 4,

  /**
   * RGBA color type.
   */
  rgba = 6,
}